requirements.txt
selenium
yagmail
undetected_chromedriver
seleniumbase
python_dotenv
slack_sdk==3.27.1